Extract editor blocks list model
